Rewrite build.gradle
Look ma, no awk or sed !

Required a bit of a thunker to realize that we don't need to parse or
read any of the emails = [...] block at this point in time. Given that
we've previously read the entire block, we know who is staying and who
is leaving, so we can just do a simple read, swap, write. Perhaps not as
cool or clever as a good awk or sed implementation but both of those
actually make it difficult to maintain the same whitespace in the file

I considered an awk solution:
`awk -v content=    emails = \["neil@winton.co", "tim@coolguy.org"\] /emails = \[/,/\]/ { print content; content=""; next } 1 build.gradle0

However this will result in several new times being written, since awk
considers each line a separate item to yield to. This reassigns content
to an empty string, but then an empty line is added, for no reasons.

I also considered implementing Niel Winton's solution:
> cat example.sed
`/emails = /,/\]/c\
emails = ["foo"]
> sed example.sed build.gradlww:xx
`